LPS-147072 create new Base class to reuse code
