+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/render/BaseFileDDMFormFieldValueRenderer.java
@@ -1,0 +1,90 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.dynamic.data.mapping.internal.render;
+
+import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
+import com.liferay.dynamic.data.mapping.model.Value;
+import com.liferay.dynamic.data.mapping.render.BaseDDMFormFieldValueRenderer;
+import com.liferay.dynamic.data.mapping.render.ValueAccessor;
+import com.liferay.dynamic.data.mapping.render.ValueAccessorException;
+import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.json.JSONException;
+import com.liferay.portal.kernel.json.JSONFactoryUtil;
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.util.Locale;
+
+/**
+ * @author Jorge García
+ */
+public abstract class BaseFileDDMFormFieldValueRenderer
+	extends	BaseDDMFormFieldValueRenderer {
+
+	@Override
+	protected ValueAccessor getValueAccessor(Locale locale) {
+		return new ValueAccessor(locale) {
+
+			@Override
+			public String get(DDMFormFieldValue ddmFormFieldValue) {
+				Value value = ddmFormFieldValue.getValue();
+
+				JSONObject jsonObject = createJSONObject(
+					value.getString(locale));
+
+				String uuid = jsonObject.getString("uuid");
+				long groupId = jsonObject.getLong("groupId");
+
+				if (Validator.isNull(uuid) && (groupId == 0)) {
+					return StringPool.BLANK;
+				}
+
+				try {
+					FileEntry fileEntry =
+						DLAppLocalServiceUtil.getFileEntryByUuidAndGroupId(
+							uuid, groupId);
+
+					return fileEntry.getTitle();
+				}
+				catch (Exception exception) {
+					if (_log.isDebugEnabled()) {
+						_log.debug(exception, exception);
+					}
+
+					return LanguageUtil.format(
+						locale, "is-temporarily-unavailable", "content");
+				}
+			}
+
+			protected JSONObject createJSONObject(String json) {
+				try {
+					return JSONFactoryUtil.createJSONObject(json);
+				}
+				catch (JSONException jsonException) {
+					throw new ValueAccessorException(jsonException);
+				}
+			}
+
+		};
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		BaseFileDDMFormFieldValueRenderer.class);
+}
