+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/render/ImageDDMFormFieldValueRenderer.java
@@ -14,85 +14,17 @@
 
 package com.liferay.dynamic.data.mapping.internal.render;
 
-import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
 import com.liferay.dynamic.data.mapping.model.DDMFormFieldType;
-import com.liferay.dynamic.data.mapping.model.Value;
-import com.liferay.dynamic.data.mapping.render.BaseDDMFormFieldValueRenderer;
-import com.liferay.dynamic.data.mapping.render.ValueAccessor;
-import com.liferay.dynamic.data.mapping.render.ValueAccessorException;
-import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.json.JSONException;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.util.Validator;
-
-import java.util.Locale;
 
 /**
- * @author Bruno Basto
- * @author Marcellus Tavares
+ * @author Jorge Garcia
  */
 public class ImageDDMFormFieldValueRenderer
-	extends BaseDDMFormFieldValueRenderer {
+	extends BaseFileDDMFormFieldValueRenderer {
 
 	@Override
 	public String getSupportedDDMFormFieldType() {
 		return DDMFormFieldType.IMAGE;
 	}
-
-	@Override
-	protected ValueAccessor getValueAccessor(Locale locale) {
-		return new ValueAccessor(locale) {
-
-			@Override
-			public String get(DDMFormFieldValue ddmFormFieldValue) {
-				Value value = ddmFormFieldValue.getValue();
-
-				JSONObject jsonObject = createJSONObject(
-					value.getString(locale));
-
-				String uuid = jsonObject.getString("uuid");
-				long groupId = jsonObject.getLong("groupId");
-
-				if (Validator.isNull(uuid) && (groupId == 0)) {
-					return StringPool.BLANK;
-				}
-
-				try {
-					FileEntry fileEntry =
-						DLAppLocalServiceUtil.getFileEntryByUuidAndGroupId(
-							uuid, groupId);
-
-					return fileEntry.getTitle();
-				}
-				catch (Exception exception) {
-					if (_log.isDebugEnabled()) {
-						_log.debug(exception, exception);
-					}
-
-					return LanguageUtil.format(
-						locale, "is-temporarily-unavailable", "content");
-				}
-			}
-
-			protected JSONObject createJSONObject(String json) {
-				try {
-					return JSONFactoryUtil.createJSONObject(json);
-				}
-				catch (JSONException jsonException) {
-					throw new ValueAccessorException(jsonException);
-				}
-			}
-
-		};
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		ImageDDMFormFieldValueRenderer.class);
 
 }
